requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
httpx
beautifulsoup4
lxml
python-multipart
aiofiles
anthropic
passlib[bcrypt]
python-jose[cryptography]
stripe
python-dotenv
bcrypt
